feat: add and fix debug output

+++ inc/CommonTypedefs.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-typedef uint32_t ID;
-typedef int32_t  TimeMoment;
+typedef uintmax_t ID;
+typedef intmax_t  TimeMoment;
 typedef bool   (*Comparator)(const void*, const void*);
 
 #endif
